+++ tests/testthat/test-rd-describe-in.R
@@ -8,7 +8,8 @@ test_that("@describeIn generic captures s3 method class", {
     f.a <- function(x) 1
   ")[[1]]
 
-  expect_equal(out$get_value("minidesc")$type, "generic")
+  expect_equal(out$get_value("minidesc")$method, TRUE)
+  expect_equal(out$get_value("minidesc")$generic, "f")
   expect_equal(out$get_value("minidesc")$label, "a")
 })
 
@@ -21,6 +22,8 @@ test_that("@describeIn generic captures s4 method class", {
     setMethod(f, signature('a'), function(x) 1)
   ")[[1]]
 
+  expect_equal(out$get_value("minidesc")$method, TRUE)
+  expect_equal(out$get_value("minidesc")$generic, "f")
   expect_equal(out$get_value("minidesc")$label, "a")
 })
 
@@ -34,6 +37,8 @@ test_that("@describeIn class captures s3 generic name", {
     mean.boo <- function(x) 1
     ")[[1]]
 
+  expect_equal(out$get_value("minidesc")$method, TRUE)
+  expect_equal(out$get_value("minidesc")$generic, "")
   expect_equal(out$get_value("minidesc")$label, "mean")
 })
 
@@ -48,37 +53,9 @@ test_that("@describeIn class captures s4 generic name", {
     setMethod('mean', 'a', function(x) 1)
     ")[[1]]
 
+  expect_equal(out$get_value("minidesc")$method, TRUE)
+  expect_equal(out$get_value("minidesc")$generic, "")
   expect_equal(out$get_value("minidesc")$label, "mean")
-})
-
-test_that("Multiple @describeIn generic combined into one", {
-  out <- roc_proc_text(rd_roclet(), "
-    #' Title
-    f <- function(x) UseMethod('f')
-
-    #' @describeIn f A
-    f.a <- function(x) 1
-
-    #' @describeIn f B
-    f.b <- function(x) 1
-  ")[[1]]
-
-  expect_equal(out$get_value("minidesc")$type, "generic")
-  expect_equal(out$get_value("minidesc")$label, c("a", "b"))
-  expect_equal(out$get_value("minidesc")$desc, c("A", "B"))
-})
-
-
-test_that("all other combinations fallback to function list", {
-  out <- roc_proc_text(rd_roclet(), "
-    #' Generic
-    foo <- function(x) UseMethod('foo')
-
-    #' @describeIn foo related function
-    bar <- function(y) {}
-  ")[[1]]
-
-  expect_equal(out$get_value("minidesc")$type, "function")
 })
 
 test_that("@describeIn class captures function name", {
@@ -90,6 +67,8 @@ test_that("@describeIn class captures function name", {
     f2 <- function(x) 1
     ")[[1]]
 
+  expect_equal(out$get_value("minidesc")$method, FALSE)
+  expect_equal(out$get_value("minidesc")$generic, "")
   expect_equal(out$get_value("minidesc")$label, "f2")
 })
 
@@ -106,6 +85,18 @@ test_that("@describeIn class captures function name with data", {
   expect_equal(out$get_value("minidesc")$label, "f2")
 })
 
+test_that("@describeIn class captures function description", {
+  out <- roc_proc_text(rd_roclet(), "
+  #' Title
+  f <- function(x) 1
+
+  #' @describeIn f A
+  f2 <- function(x) 1
+  ")[[1]]
+
+  expect_equal(out$get_value("minidesc")$desc, "A")
+})
+
 test_that("function names are escaped", {
   out <- roc_proc_text(rd_roclet(), "
     #' foo
@@ -114,5 +105,58 @@ test_that("function names are escaped", {
     #' @describeIn foo shortcut for foo
     `%foo%` <- function(x, y) foo(x, y)
     ")[[1]]
+  out
   expect_match(out$get_rd("minidesc"), "\\\\%foo\\\\%")
 })
+
+test_that("minidescription is correctly formatted", {
+  x <- rd_section_minidesc(
+    df = subsection(label = c("foo", "bar"), desc = c("baz", "qux")),
+    method = TRUE,
+    generic = "corge"
+  )
+  expect_equal_strings(
+    format(x),
+    "\\section{Related Functions and Methods}{\n
+    \\subsection{Methods extending \\code{corge} generic (by class):}{
+    \n\\itemize{\n\\item \\code{foo}: baz\n\\item \\code{bar}: qux\n}\n
+    }\n}\n"
+  )
+})
+
+test_that("Multiple @describeIn functions combined into one", {
+  example <- test_path("roxygen-block-describe-in-functions.R")
+  out <- roc_proc_text(rd_roclet(), brio::read_file(example))[[1]]
+
+  expect_equal(out$get_value("minidesc")$method, c(FALSE, FALSE))
+  expect_equal(out$get_value("minidesc")$generic, c("", ""))
+  expect_equal(out$get_value("minidesc")$label, c("square", "cube"))
+})
+
+test_that(
+  "Multiple @describeIn methods and others are combined into a generic", {
+    example <- test_path("roxygen-block-describe-in-method-in-gen.R")
+    out <- roc_proc_text(rd_roclet(), brio::read_file(example))[[1]]
+    expect_equal(out$get_value("minidesc")$method, c(TRUE, TRUE, FALSE, FALSE))
+    expect_equal(out$get_value("minidesc")$generic, c("zap", "zap", "", ""))
+    expect_equal(
+      out$get_value("minidesc")$label,
+      c("numeric", "character", "print.qux", "zap_helper")
+    )
+  }
+)
+
+test_that(
+  "Multiple @describeIn methods and others are combined into a class constructor",
+  {
+    example <- test_path("roxygen-block-describe-in-method-in-const.R")
+    out <- roc_proc_text(rd_roclet(), brio::read_file(example))[[1]]
+
+    expect_equal(out$get_value("minidesc")$method, c(TRUE, TRUE, FALSE, FALSE))
+    expect_equal(out$get_value("minidesc")$generic, rep("", 4))
+    expect_equal(
+      out$get_value("minidesc")$label, 
+      c("print", "format", "format.bar", "is_foo")
+    )
+  }
+)